+++ Flow/src/utopia/flow/view/immutable/caching/Lazy.scala
@@ -1,7 +1,8 @@
 package utopia.flow.view.immutable.caching
 
 import utopia.flow.collection.immutable.caching.iterable.LazySeq
-import utopia.flow.collection.mutable.iterator.{LazyInitIterator, PollableOnce, RepeatOneForeverIterator}
+import utopia.flow.collection.mutable.iterator.{LazyInitIterator, PollableOnce}
+import utopia.flow.collection.CollectionExtensions._
 import utopia.flow.util.logging.Logger
 import utopia.flow.view.immutable.View
 import utopia.flow.view.immutable.eventful.ListenableLazy
@@ -56,6 +57,36 @@ object Lazy
 	  * @return A no-op lazy that simply wraps the value
 	  */
 	def initialized[A](value: A) = PreInitializedLazy(value)
+	/**
+	  * @param options Available lazily initialized containers
+	  * @param cache Whether caching is allowed.
+	  *              If true, once a value is acquired, it is cached and returned on all subsequent calls to value.
+	  *              If false (default), the value may change based on how the containers become available.
+	  * @tparam A Type of the wrapped / yielded value
+	  * @return A lazy container selecting from one of the specified options
+	  */
+	// If only one container is specified, yields that
+	def anyOf[A](options: Seq[Lazy[A]], cache: Boolean = false) = options.only.getOrElse {
+		// Case: Caching enabled => Checks whether a value is immediately available
+		if (cache)
+			options.findMap { _.current } match {
+				case Some(value) => initialized(value)
+				case None => new FirstAvailableLazy(options)
+			}
+		else
+			new PrioritizingLazy(options)
+	}
+	
+	
+	// EXTENSIONS   --------------------
+	
+	implicit class DeepLazy[+A](val l: Lazy[Lazy[A]]) extends AnyVal
+	{
+		/**
+		  * @return A flattened version of this lazy container
+		  */
+		def flatten = new FlatteningLazy[A](l)
+	}
 	
 	
 	// NESTED   ------------------------
@@ -127,7 +158,7 @@ trait Lazy[+A] extends View[A]
 	  * @tparam B Type of mapping result
 	  * @return Lazily initialized mapping results
 	  */
-	def map[B](f: A => B) = Lazy { f(value) }
+	def map[B](f: A => B): Lazy[B] = Lazy { f(value) }
 	/**
 	  * Maps the contents of this container.
 	  * If the current value has already been calculated, performs the
@@ -138,15 +169,55 @@ trait Lazy[+A] extends View[A]
 	  * @return A new possibly lazily initialized container that contains
 	  *         the mapping results.
 	  */
-	def mapCurrent[B](f: A => B) = current match {
+	@deprecated("Renamed to .lightMap(...)", "v2.7")
+	def mapCurrent[B](f: A => B) = lightMap(f)
+	/**
+	  * Maps the contents of this container.
+	  * Unlike [[map]], the mapping function call is not wrapped in a Lazy by itself and may be called immediately.
+	  * @param f A mapping function to perform
+	  * @tparam B Type of mapping results
+	  * @return A lazily initialized container that has [[current]] as soon as this container has,
+	  *         but yields a mapped value.
+	  */
+	def lightMap[B](f: A => B): Lazy[B] = current match {
 		case Some(value) => Lazy.initialized(f(value))
-		case None => map(f)
+		case None => new MappingLazyView(this)(f)
 	}
+	/**
+	  * @param f A mapping function that yields a lazily acquired result
+	  * @tparam B Type of mapping results, once acquired
+	  * @return A lazily initialized container that contains the result of 'f'
+	  */
+	def flatMap[B](f: A => Lazy[B]): Lazy[B] = new FlatteningLazy(lightMap(f))
+	
 	/**
 	  * Lazily maps this container's content
 	  * @param f A mapping function applied lazily. Yields 0-n values (via iterator).
 	  * @tparam B Type of individual map results
 	  * @return A lazily initialized collection based on the mapping results
 	  */
-	def flatMap[B](f: A => IterableOnce[B]) = LazySeq.from[B](LazyInitIterator { f(value) })
+	def mapToSeq[B](f: A => IterableOnce[B]) = LazySeq.from[B](LazyInitIterator { f(value) })
+	
+	/**
+	  * Combines this with another lazy container, taking the first available value
+	  * @param other Another lazy container
+	  * @param cache Whether the acquired value should be cached and kept the same.
+	  *              If false (default), the returned value may change to that of this lazy container
+	  *              in instances where 'other' is first initialized and this becomes initialized later.
+	  * @tparam B Type of values acquired from the resulting container
+	  * @return A lazy container that takes its value from either of these containers
+	  */
+	def or[B >: A](other: Lazy[B], cache: Boolean = false) = {
+		if (cache)
+			current.orElse { other.current } match {
+				case Some(value) => Lazy.initialized(value)
+				case None => FirstAvailableLazy(this, other)
+			}
+		else
+			PrioritizingLazy(this, other)
+	}
+	/**
+	  * An alias for [[or]]
+	  */
+	def ||[B >: A](other: Lazy[B]) = or(other)
 }